[3-4단계 방탈출 예약 대기] 포포(이민우) 미션 제출합니다. (#275)
* refactor: 예약 가능 시간 조회 API Endpoint 수정

* refactor: 인기 테마 조회를 예약 일급컬렉션을 이용하도록 수정

* refactor: 시간별 예약 가능 여부를 예약 일급컬렉션을 이용하도록 수정

* refactor: 예약 객체 생성자 변경

* refactor: 예약 슬롯 객체 분리

* refactor: 예약 일시 객체를 슬롯에 병합

* feat: 예약 대기열 객체 구현

* feat: 중복 예약 대기 예외처리

* feat: 예약 대기 서비스 구현

* feat: 예약 대기 컨트롤러 구현

* refactor: 예약 상태 응답 변경

* refactor: repository 수정 메서드 @Transactional 추가

* refactor: 예약 삭제를 어드민만 실행할 수 있도록 변경

* feat: 예약 대기 취소 서비스 구현

* feat: 예약 대기 취소 컨트롤러 구현

* fix: 다른 사용자의 예약 대기를 취소할 수 없도록 수정

* feat: 예약 대기상태인 지 확인하는 메서드 추가

* refactor: 서비스 응답 Waiting 객체 분리

* refactor: 예약 주체를 사용자로 변경

* refactor: 예약 슬롯 별 대기열 객체 분리

* chore: 프로덕션 DB URL 변경

* feat: 모든 대기 예약 조회 구현

* refactor: 예약 서비스 메서드 분리

* test: 예약 취소 기능 테스트

* feat: 어드민 예약 대기 거절 구현

* docs: 예약 대기 기능

* refactor: 서비스 테스트 환경 설정

서비스 테스트들에 대해 아래와 같이 적용합니다.
- Spring Data JPA 빈은 로드합니다.
- 테스트 클래스, 메서드의 트랜잭션을 끕니다.
- 매 테스트 이후 DB 데이터를 초기화합니다.

* fix: 확정된 예약 삭제 시 첫번째 대기가 확정되도록 수정

* refactor: repository가 일급 컬렉션을 반환하도록 변경

* refactor: repository 공통 기능 일반화

* refactor: repository 인프라 계층 분리

* refactor: Specification 중복 코드 제거

* refactor: 특정 기능에 종속적인 서비스 메서드 네이밍 변경

* refactor: 테이블 명 대문자로 통일

* refactor: 대기 순서를 id가 아닌 예약 대기 시간으로 비교하도록 변경

* refactor: repository 메서드 파라미터명 변경

* style: 불필요한 괄호 제거

* fix: 대기 순서를 예약 대기 시작 시간으로 비교

* refactor: 불필요한 메서드 제거

* refactor: 테스트 클래스명 변경

* refactor: Reservation Slot 네이밍 변경

* refactor: RoomescapeSchedule 패키지 이동

* refactor: 예약 스케줄과 예약 일시 분리

* refactor: 변경되지 않은 예약 스케줄 네이밍 변경

* refactor: 가독성 향상을 위한 메서드 분리

* refactor: 예약에서 시간, 테마에 바로 접근할 수 있도록 수정

+++ src/main/java/roomescape/presentation/rest/ReservationController.java
@@ -19,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.ReservationService;
 import roomescape.domain.auth.AuthenticationInfo;
 import roomescape.domain.reservation.ReservationSearchFilter;
+import roomescape.exception.AuthorizationException;
 import roomescape.presentation.request.CreateReservationRequest;
 import roomescape.presentation.response.ReservationResponse;
 
@@ -39,6 +40,16 @@ public class ReservationController {
         return ReservationResponse.from(reservation);
     }
 
+    @PostMapping("/wait")
+    @ResponseStatus(CREATED)
+    public ReservationResponse waitFor(
+            final AuthenticationInfo authenticationInfo,
+            @RequestBody @Valid final CreateReservationRequest request
+    ) {
+        var reservation = service.waitFor(authenticationInfo.id(), request.date(), request.timeId(), request.themeId());
+        return ReservationResponse.from(reservation);
+    }
+
     @GetMapping
     public List<ReservationResponse> getAllReservations(
             @RequestParam(name = "themeId", required = false) final Long themeId,
@@ -53,7 +64,23 @@ public class ReservationController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(NO_CONTENT)
-    public void delete(@PathVariable("id") final long id) {
+    public void delete(
+        final AuthenticationInfo authenticationInfo,
+        @PathVariable("id") final long id
+    ) {
+        if (!authenticationInfo.isAdmin()) {
+            throw new AuthorizationException("관리자에게만 허용된 작업입니다.");
+        }
         service.removeById(id);
+    }
+
+    @DeleteMapping("/wait/{id}")
+    @ResponseStatus(NO_CONTENT)
+    public void cancelWaiting(
+        final AuthenticationInfo authenticationInfo,
+        @PathVariable("id") final long reservationId
+    ) {
+        var userId = authenticationInfo.id();
+        service.cancelWaiting(userId, reservationId);
     }
 }
